requirements.txt
python-telegram-bot==20.0
psycopg2==2.9.6
